+++ packages/core/src/types/ContourData.ts
@@ -8,12 +8,14 @@ type ContourSetData = {
   data: ContourData[];
   frameOfReferenceUID: string;
   color?: Point3;
+  segmentIndex?: number;
 };
 
 type ContourData = {
   points: Point3[];
   type: ContourType;
-  color?: Point3;
+  color: Point3;
+  segmentIndex: number;
 };
 
 export type { PublicContourSetData, ContourSetData, ContourData };