+++ packages/core/src/types/IGeometry.ts
@@ -1,11 +1,11 @@
-import { ContourSet } from '../cache/classes/ContourSet';
 import { GeometryType } from '../enums';
+import { IContourSet } from './IContourSet';
 
 // interface IGeometry can be array of IContourSet
 interface IGeometry {
   id: string;
   type: GeometryType;
-  data: ContourSet; // | Array<IClosedSurface> , etc
+  data: IContourSet; // | Array<IClosedSurface> , etc
   sizeInBytes: number;
 }