+++ packages/tools/src/tools/displayTools/Contour/contourConfig.ts
@@ -1,0 +1,15 @@
+import { ContourConfig } from '../../../types/ContourTypes';
+
+const defaultContourConfig: ContourConfig = {
+  renderOutline: true,
+  outlineWidthActive: 2,
+  outlineWidthInactive: 2,
+  outlineOpacity: 1,
+  outlineOpacityInactive: 0.85,
+};
+
+function getDefaultContourConfig(): ContourConfig {
+  return defaultContourConfig;
+}
+
+export default getDefaultContourConfig;